feat: :art: Improved file structure and separation of home page
Moved nested components to designated folder and home component to its own subdirectory
