+++ src/routes/index.tsx
@@ -1,5 +1,5 @@
 import { createFileRoute } from "@tanstack/react-router";
-import Home from "../components/Home";
+import Home from "../components/home/Home";
 
 export const Route = createFileRoute("/")({
   component: Home,
